moved the data related code to a data/ directory.
