+++ backend/src/data/dataStore.ts
@@ -1,5 +1,5 @@
 import fs from 'fs';
-import { DataStore, Session, SessionStore } from '../../shared/interfaces';
+import { DataStore, Session, SessionStore } from '../../../shared/interfaces';
 import { v4 as uuidv4 } from 'uuid';
 import jwt from 'jsonwebtoken';
 import crypto from 'crypto';
@@ -10,8 +10,8 @@ import crypto from 'crypto';
 let sessionStore: SessionStore = { sessions: [] };
 let database: DataStore = { users: [], elections: [] };
 
-const SESSION_PATH = "./src/sessions.json";
-const DATA_PATH = "./src/database.json";
+const SESSION_PATH = "./src/data/sessions.json";
+const DATA_PATH = "./src/data/database.json";
 
 // TODO: move this to a .env file! and put the .env in .gitignore
 const secretKey = 'abcde12345';
@@ -101,7 +101,7 @@ export function setData(newData: DataStore) {
 // clears the entire database, as well as clears out all existing sessions.
 export function clear() {
   // Reset both memory and files
-  setData({ users: [] });
+  setData({ users: [], elections: [] });
   setSessions([]);
 
   return {};
