+++ backend/src/data/database.json
@@ -1,0 +1,4 @@
+{
+  "users": [],
+  "elections": []
+}
