--- backend/src/database.json
@@ -1,3 +1,0 @@
-{
-  "users": []
-}
